add temp create_book method to see if it works(it does ;))

+++ migrations/2023-07-17-093014_create_books/up.sql
@@ -4,6 +4,6 @@ CREATE TABLE books (
 	title VARCHAR NOT NULL,
 	author VARCHAR NOT NULL,
 	genre VARCHAR NOT NULL,
-	rating REAL NOT NULL,
+	rating INTEGER NOT NULL,
 	status VARCHAR NOT NULL
 )
